backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
PyJWT==2.9.0
cryptography==43.0.0
httpx==0.27.0
python-multipart==0.0.9
structlog==24.4.0
supabase==2.9.0

